todo: create sink table
